+++ controllers/indexController.js
@@ -1,6 +1,6 @@
 // screenController
 var indexController = function(req, res, next) {
-	
+	// TODO: move this to utils.js
 	var generateHash = function() {
 	    var text = "";
 	    var possible = "ABCDEFGHIJKLMNOPQRSTUVWXYZabcdefghijklmnopqrstuvwxyz0123456789";
@@ -11,7 +11,9 @@ var indexController = function(req, res, next) {
 	    return text;
 	};
 
-	res.redirect('/screen/'+generateHash());
+	var hash = generateHash();
+
+	res.redirect('/screen/'+hash);
 };
 
 module.exports = indexController;
